third_party/WebKit/LayoutTests/imported/web-platform-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/load-removes-queued-error-event-expected.txt

This is a testharness.js-based test.
FAIL video error event assert_array_equals: property 1, expected "loadstart" but got "error"
FAIL source error event assert_array_equals: property 1, expected "loadstart" but got "error"
Harness: the test ran to completion.
